Show warning when parameter name is parsed incorrectly
AWS credentials are parsed by splittling each key-value pair on the
_last_ `=` character. As a result, if the value contains an `=`
character then both the key and the value are parsed incorrectly.
However, the incorrectly parsed key is only compared with the expected
key name up until the length of the latter, and is therefore seen as a
valid parameter which is assigned the incorrectly parsed value.

For example, passing a parameter of `"session-token=/822koPQs="` results
in a parameter name of `"session-token=/822koPQs"` and an empty value.
Since the first 13 characters match `"session-token"` then this empty
value is accepted and assigned to the session token.

Expose the parsing error by checking the entire length of the parameter
name, therefore resulting in an "unknown parameter" warning in the above
scenario.