Fix parsing of credential parameters
AWS credentials are passed as key-value pairs in the form of
`"key=value"`. The key and the value are extracted by splitting the
string on the _last_ occurance of `"="`. This means that if the value
contains an `"="` then both the key name and the value are parsed
incorrectly. For example `"session-token=woirInBw="` results in a key of
`"session-token=woirInBw"` and an empty value, which now results in an
"unknown parameter" warning. Fix this, and maintain backwards
compatibility, by splitting on the _first_ occurence of `"="` rather
than the last.