tests: updated query_seeking to actually test the element
The GstElement base class randomly selects pads when querying
and then hands off the request to the src pad or, if it's a
sink pad, gets the peer to that pad and queries that instead.
So this test was not actually testing the element's own
responses to the queries because, having no peer, the query
would get dropped in this test, given that the element is
not connected to anything.

This patch directly queries this sink element's sink pad
and verifies the expected results w.r.t., the seek behavior.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>