tests: added the start of a 'header rewrite' test
This test pushes a couple buffers worth of packets, then
sends a segment.pos = 0 (seek to head) followed by another
packet as if writing over the header.  In this sequence,
we're expecting to see the upload part count stay the same
until the eos causes a flush to actually write the header.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>