s3sink: exposed constructors for uploader and downloader on class
The s3sink implementation destroys the uploader instance and
recreates it as needed during flushing events (end of stream,
seek, etc.) which makes trying to test longer sequences of
behaviors like seeking much harder.  While we're here, I'm
exposing the downloder too since that is a black box to the
test environment.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>