s3sink: reworked the GstS3Downloader to be test-friendly
The class and files have been re-written to mimic the
design pattern used by the GstS3Uploader and the derived
implementation class, GstS3MultipartUploader.  This is
so we can similarly create a class for the downloader
in the test environment and perform tests that expect a
download operation to work.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>