s3sink-cached test clean-up
The tests don't need to re-test the upload/download/complete
fail conditions in the s3sink.c, so I've removed those
fields for clarity.  I've also added a header, and clarified
that the stringstream buffer in the previous commit is
intended to represent the uploaded buffer on AWS as it's
populated by a test.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>