s3sink-cached: added test for forward cache boundary
Modified the test apparatus further to represent the
uploader's view of the not-completed buffer vs. the
buffer that can be downloaded from to do other
completion operations when crossing cache boundaries.

Added a test for when seeking back into the cache and
then continuing to write until crossing into a cache-
miss situation where the uploader state should be
considered dirty (above) and requiring completion.

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>