video-2456: exposed sdk retry, scale as properties
Added a test to prove out the change in behavior.
The original behavior resulted in 25.550 seconds of
delay going from NULL to READY if anything impacting
the URL was misconfigured because each of the 10
fruitless attempts would successively and exponentially
"back off" 0, 50, 100, 200, etc. milliseconds until the
10th attempt.  By  being allowed to reduce the attempts
to say 3 for example, that total delay can be reduced
to something more manageable (150ms).

Signed-off-by: Thomas Goodwin <thomas.goodwin@laerdal.com>